pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Page Analyzer"

dependencies = [
    "flask==3.0.0",
    "python-dotenv==1.0.0",
    "gunicorn==21.2.0",
    "ruff==0.1.0",
    "psycopg2-binary==2.9.9",
    "validators==0.22.0",
    "requests==2.31.0",
    "beautifulsoup4==4.12.2",
]
requires-python = ">=3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["page_analyzer"]